kernel: don't check ND_REGION_PAGEMAP
DAX is disabled because the kernel doesn't know the namespace
information of the DAX device (kata-containers.img). As a temporal solution
to this issue, kernel shouldn't check the ND_REGION_PAGEMAP bit. The permanent
solution will be to write the metadata that kernel needs directly in the
kata-containers.img but this will require time since currently there are not
tools to this.

fixes #378

Signed-off-by: Julio Montes <julio.montes@intel.com>