kata-deploy: consider stateless configuration path
Some distros like Clearlinux may not have the config files
under `/etc` as they use `/usr/share/defaults` as their
stateless configuration path, so we need to take this path
in consideration.

This change will first verify if config file exists in `/etc`
and if not found, it use the config file in the stateless path.

Fixes: #833.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>