qemu: Fix experimental virtio-fs crash on RHCOS
This will fix a failure of experimental virtiofsd trying to mount a
directory on temporary filesystem that might be read-only. It is the
case in RHCOS which has virtiofsd crashing with "operation not permitted".

Cherry-picked the following commit from QEMU:
 * virtiofsd: avoid /proc/self/fd tempdir (commit ebf101955ce8f8d72f)

Note: this fix won't be applied on our build of the upstream qemu.

Fixes #1176

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>